# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Football Kickoff 2026 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SignupInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthUser" }
...
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyOtpInput" }
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleOk" }
...
  /tournaments/{slug}/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [leaderboard]
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
        - in: query
          name: filter
          required: false
          schema: { type: string, enum: [overall, group_stage, knockouts, this_week] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LeaderboardEntry" }
...
components:
  schemas:
...
    SignupInput:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
    AuthUser:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string, enum: [user, admin] }
        createdAt: { type: string, format: date-time }
...
    LeaderboardEntry:
      type: object
      required: [userId, displayName, rank, totalPoints, exactScores, correctResults, goalDifferenceHits, oneTeamScoreHits, lastMatchPoints, isMe]
      properties:
        userId: { type: integer }
        displayName: { type: string }
        rank: { type: integer }
        totalPoints: { type: number }
        exactScores: { type: integer }
        correctResults: { type: integer }
        goalDifferenceHits: { type: integer }
        oneTeamScoreHits: { type: integer }
        lastMatchPoints: { type: number }
        isMe: { type: boolean }
...
    SimpleOk:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    VerifyOtpInput:
      type: object
      required: [code]
      properties:
        code: { type: string, minLength: 6, maxLength: 6 }
...